+++ src/helpers/constants.ts
@@ -1,5 +1,6 @@
 export const config = {
   defaultPriority: ['theme-variable', 'variable', 'mixin'],
   sassImportRegex: /@import ['"]([^'"]+)['"];/g,
+  parseFilePathRegex: /^.*[\\\/]/,
   fileType: 'scss',
 };