+++ src/helpers/parse-param.ts
@@ -1,1 +1,5 @@
-export class ParseParam {}
+// export class Params {
+//   public bundleParam;
+//   constructor() {}
+//   public readParamFromFile() {}
+// }