+++ src/index.ts
@@ -1,104 +1,56 @@
-import { writeFile, statSync, readFileSync } from 'fs';
-import { resolve, join, dirname } from 'path';
+import { readFileSync } from 'fs';
+import { join } from 'path';
 import * as globby from 'globby';
 import { StringDecoder } from 'string_decoder';
-import { config } from './helpers/constants';
 import { Sort } from './helpers/Sort';
+import { cwDir, resolveDirDest, writeAsync } from './helpers/fs-utils';
+import { getUniqueScss, removeImports } from './helpers/file-content-utils';
+import { logger } from './helpers/logger';
+import { Params } from './helpers/Params';
+import { IParams } from './interface/IParams';
 
 const decoder = new StringDecoder('utf8');
-const log = (info: Error | string) => console.log(info);
-const isFile = (f: string) => statSync(f).isFile();
-
-const removeImports = (content: string) =>
-  content.replace(config.sassImportRegex, '');
-
-const readSync = (filePath: string) => readFileSync(filePath, 'utf8');
-const getUniqueScss = (files: Array<string>) => {
-  const scssImports = files
-    .map(file => {
-      let baseDir = dirname(file);
-      return getImports(readSync(file), baseDir);
-    })
-    .reduce((acc, curr) => acc.concat(curr), []);
-  const allImports = [...scssImports, ...files];
-
-  return [...new Set(allImports)];
-};
-
-const getImports = (
-  content: string,
-  baseDir: string,
-  imports: Array<string> = []
-) => {
-  let match;
-  while ((match = config.sassImportRegex.exec(content)) !== null) {
-    const pathFile = defineExtension(join(baseDir, match[1]));
-    if (!imports.some(el => el === pathFile)) {
-      imports.push(pathFile);
-      getImports(readSync(pathFile), dirname(pathFile), imports);
-    }
-  }
-
-  return imports;
-};
-
-const writeAsync = (path: string, content: string) => {
-  return new Promise((res, rej) => {
-    writeFile(resolve(path), content, error => {
-      if (error) {
-        return rej(error);
-      }
-
-      return res(content);
-    });
-  });
-};
-
-const defineExtension = (filePath: string) => {
-  const justScss = `${filePath}.${config.fileType}`;
-
-  if (isFile(justScss)) {
-    return justScss;
-  }
-  throw new Error(`⛔ ⛔ ⛔ No file for module ${filePath}`);
-};
 
 export = (
   mask: string[] | string,
   dest: string,
-  sort: string[] = config.defaultPriority
+  sort: string[] | string,
+  config: boolean
 ) => {
-  const fullPath = join(process.cwd());
-  const sortOrder = Array.isArray(sort) ? sort : [sort];
-  const sortInstance = new Sort(sortOrder);
-  if (!mask || !mask.length) {
+  const params: IParams = new Params(mask, dest, sort, config).param;
+
+  if (!params.mask || !params.mask.length) {
     throw new Error('⛔ ⛔ ⛔ Please provide the src for concat method');
   }
-  const searchMask = Array.isArray(mask) ? mask : [mask];
+
+  resolveDirDest(params.dest);
+
+  const searchMask = Array.isArray(params.mask) ? params.mask : [params.mask];
+
   return globby(searchMask).then(paths => {
-    const files = paths.map(file => join(fullPath, file));
+    const files = paths.map(file => join(cwDir(), file));
 
     const unique = getUniqueScss(files);
-    const sorted = sortInstance.sort(unique);
+    const sorted = new Sort(params.sort).sort(unique);
     const buffers = sorted.map(file => {
       return readFileSync(file);
     });
     const buff = Buffer.concat(buffers);
     let utfFormat = decoder.write(buff);
 
-    if (dest) {
-      log(`⏳ ⏳ ⏳ Saving result to ${dest}...`);
+    if (params.dest) {
+      logger(`⏳ ⏳ ⏳ Saving result to ${params.dest}  ...`);
       const utf = removeImports(utfFormat);
-      return writeAsync(dest, utf)
+      return writeAsync(params.dest, utf)
         .then(() => {
-          log(`🚀 🚀 🚀 SAVED SUCCESSFULLY \nPlease check ${dest}`);
+          logger(`🚀 🚀 🚀 SAVED SUCCESSFULLY \nPlease check ${params.dest}`);
           return utf;
         })
         .catch(reason => {
-          log(`⛔ ⛔ ⛔\n${reason}`);
+          logger(`⛔ ⛔ ⛔\n${reason}`);
         });
     }
-    log('📁 Please provide destination option ');
+    logger('📁 Please provide destination option ');
     throw new Error('📁 Please provide destination option ');
   });
 };