+++ src/interface/IParams.ts
@@ -1,0 +1,5 @@
+export interface IParams {
+  mask: Array<string> | string;
+  dest: string;
+  sort?: Array<string> | string;
+}