+++ CHANGELOG.md
@@ -1,9 +1,9 @@
 # 1.1.0  (2018-03-09)
 
- * code refactor
+* code refactor
 
 ### Feature
- * ???
+* Added vulnerabilities testing
 
 # 1.0.2  (2018-03-06)